add the typedef for ulong and fix a few definition issues for imac.

+++ sz/src/sz_double.c
@@ -193,7 +193,7 @@ void SZ_compress_args_double_NoCkRngeNoGzip_2D(char** newByteData, double *oriDa
 	int i,j;
 	double pred1D, pred2D;
 	double diff = 0.0;
-	long itvNum = 0;
+	double itvNum = 0;
 	double *P0, *P1;
 		
 	int dataLength = r1*r2;	
@@ -264,7 +264,7 @@ void SZ_compress_args_double_NoCkRngeNoGzip_2D(char** newByteData, double *oriDa
 	pred1D = P1[0];
 	diff = spaceFillingValue[1] - pred1D;
 
-	itvNum = (long) (fabs(diff)/realPrecision) + 1;
+	itvNum =  fabs(diff)/realPrecision + 1;
 
 	if (itvNum < 256)
 	{
@@ -298,7 +298,7 @@ void SZ_compress_args_double_NoCkRngeNoGzip_2D(char** newByteData, double *oriDa
 		pred1D = 2*P1[j-1] - P1[j-2];
 		diff = spaceFillingValue[j] - pred1D;
 
-		itvNum = (long) (fabs(diff)/realPrecision) + 1;
+		itvNum = fabs(diff)/realPrecision + 1;
 
 		if (itvNum < 256)
 		{
@@ -336,7 +336,7 @@ void SZ_compress_args_double_NoCkRngeNoGzip_2D(char** newByteData, double *oriDa
 		pred1D = P1[0];
 		diff = spaceFillingValue[index] - pred1D;
 
-		itvNum = (long) (fabs(diff)/realPrecision) + 1;
+		itvNum = fabs(diff)/realPrecision + 1;
 
 		if (itvNum < 256)
 		{
@@ -372,7 +372,7 @@ void SZ_compress_args_double_NoCkRngeNoGzip_2D(char** newByteData, double *oriDa
 
 			diff = spaceFillingValue[index] - pred2D;
 
-			itvNum = (long) (fabs(diff)/realPrecision) + 1;
+			itvNum = fabs(diff)/realPrecision + 1;
 
 			if (itvNum < 256)
 			{
@@ -453,7 +453,7 @@ void SZ_compress_args_double_NoCkRngeNoGzip_3D(char** newByteData, double *oriDa
 	int i,j,k;
 	double pred1D, pred2D, pred3D;
 	double diff = 0.0;
-	long itvNum = 0;
+	double itvNum = 0;
 	double *P0, *P1;
 
 	int dataLength = r1*r2*r3;
@@ -526,7 +526,7 @@ void SZ_compress_args_double_NoCkRngeNoGzip_3D(char** newByteData, double *oriDa
 	pred1D = P1[0];
 	diff = spaceFillingValue[1] - pred1D;
 
-	itvNum = (long) (fabs(diff)/realPrecision) + 1;
+	itvNum = fabs(diff)/realPrecision + 1;
 
 	if (itvNum < 256)
 	{
@@ -560,7 +560,7 @@ void SZ_compress_args_double_NoCkRngeNoGzip_3D(char** newByteData, double *oriDa
 		pred1D = 2*P1[j-1] - P1[j-2];
 		diff = spaceFillingValue[j] - pred1D;
 
-		itvNum = (long) (fabs(diff)/realPrecision) + 1;
+		itvNum = fabs(diff)/realPrecision + 1;
 
 		if (itvNum < 256)
 		{
@@ -598,7 +598,7 @@ void SZ_compress_args_double_NoCkRngeNoGzip_3D(char** newByteData, double *oriDa
 		pred1D = P1[index-r3];
 		diff = spaceFillingValue[index] - pred1D;
 
-		itvNum = (long) (fabs(diff)/realPrecision) + 1;
+		itvNum = fabs(diff)/realPrecision + 1;
 
 		if (itvNum < 256)
 		{
@@ -634,7 +634,7 @@ void SZ_compress_args_double_NoCkRngeNoGzip_3D(char** newByteData, double *oriDa
 
 			diff = spaceFillingValue[index] - pred2D;
 
-			itvNum = (long) (fabs(diff)/realPrecision) + 1;
+			itvNum = fabs(diff)/realPrecision + 1;
 
 			if (itvNum < 256)
 			{
@@ -674,7 +674,7 @@ void SZ_compress_args_double_NoCkRngeNoGzip_3D(char** newByteData, double *oriDa
 		pred1D = P1[0];
 		diff = spaceFillingValue[index] - pred1D;
 
-		itvNum = (long) (fabs(diff)/realPrecision) + 1;
+		itvNum = fabs(diff)/realPrecision + 1;
 
 		if (itvNum < 256)
 		{
@@ -710,7 +710,7 @@ void SZ_compress_args_double_NoCkRngeNoGzip_3D(char** newByteData, double *oriDa
 			pred2D = P0[j-1] + P1[j] - P1[j-1];
 			diff = spaceFillingValue[index] - pred2D;
 
-			itvNum = (long) (fabs(diff)/realPrecision) + 1;
+			itvNum = fabs(diff)/realPrecision + 1;
 
 			if (itvNum < 256)
 			{
@@ -749,7 +749,7 @@ void SZ_compress_args_double_NoCkRngeNoGzip_3D(char** newByteData, double *oriDa
 			pred2D = P0[index2D-r3] + P1[index2D] - P1[index2D-r3];
 			diff = spaceFillingValue[index] - pred2D;
 
-			itvNum = (long) (fabs(diff)/realPrecision) + 1;
+			itvNum = fabs(diff)/realPrecision + 1;
 
 			if (itvNum < 256)
 			{
@@ -785,7 +785,7 @@ void SZ_compress_args_double_NoCkRngeNoGzip_3D(char** newByteData, double *oriDa
 				pred3D = P0[index2D-1] + P0[index2D-r3]+ P1[index2D] - P0[index2D-r3-1] - P1[index2D-r3] - P1[index2D-1] + P1[index2D-r3-1];
 				diff = spaceFillingValue[index] - pred3D;
 
-				itvNum = (long) (fabs(diff)/realPrecision) + 1;
+				itvNum = fabs(diff)/realPrecision + 1;
 
 				if (itvNum < 256)
 				{
